Added selection of publication protocols via configuration Added HTTP protocol update via A and AAAA records Added Removal DNS entries when protocol disabled DNS updates only if value of entry changes

+++ pinning-services/index.js
@@ -1,6 +1,8 @@
 const cron = require('cron');
 const fetch = require('node-fetch');
 const fs = require('fs');
+const httpPublisherIpv4Address = "198.50.215.6";
+const httpPublisherIpv6Address = "2607:5300:203:4014:5074:f2ff:feb1:a87f";
 const hyperPublisher = require('hyperdrive-publisher')
 const ipfsClient = require('ipfs-http-client')
 
@@ -81,73 +83,121 @@ const job = new cron.CronJob(period, function() {
         const dirWww = `${projDir}/${projName}/www`;
         const dirApi = `${projDir}/${projName}/api`;
 
-        // Pin WWW site to Hypercore
-        if (fs.existsSync(dirWww)) {
-          getDatSeed('dat-seed-www', projName, domain, txtHypercoreWww)
-            .then(seed => hyperPublisher.sync({
-              seed,
-              fsPath: dirWww,
-              drivePath: '/',
-              syncTime: 600000 // Allow up to 10 minutes for sync
-            }))
-            .then(({diff, url}) => {
-              // Log any changes to hyperdrive and return url
-              console.log(`WWW site for ${projName} pinned at ${url}. Changes:`);
-              console.log(diff);
-              return url;
-            })
-            .catch(function(error) {
-              console.log(error);
+        //Read publication settings from project configuration
+        let protoHttp = true;
+        let protoHttpDnsUpdate = true;
+        let protoIpfs = true;
+        let protoHypercore = true;
+        if (proj['publication']) {
+          if (proj['publication']['protocol']) {
+
+            // Check for known protocols and set flags
+            Object.keys(proj['publication']['protocol']).forEach((item) => {
+              switch (item) {
+                case "http":
+                  protoHttp = (proj['publication']['protocol'][item]["enable"] === true) ? true : false;
+                  protoHttpDnsUpdate = (proj['publication']['protocol'][item]["dnsupdate"] === true) ? true : false;
+                  break;
+                case "ipfs":
+                  protoIpfs = (proj['publication']['protocol'][item]["enable"] === true) ? true : false;
+                  break;
+                case "hypercore":
+                  protoHypercore = (proj['publication']['protocol'][item]["enable"] === true) ? true : false;
+                  break;
+                default:
+                  console.log(`WARNING: Unknown protocol in publication project JSON - ${item}`);
+              }
             });
+          }
+        }
+
+        // Pin WWW site to Hypercore
+        if (protoHypercore) {
+          if (fs.existsSync(dirWww)) {
+            getDatSeed('dat-seed-www', projName, domain, txtHypercoreWww)
+              .then(seed => hyperPublisher.sync({
+                seed,
+                fsPath: dirWww,
+                drivePath: '/',
+                syncTime: 600000 // Allow up to 10 minutes for sync
+              }))
+              .then(({ diff, url }) => {
+                // Log any changes to hyperdrive and return url
+                console.log(`WWW site for ${projName} pinned at ${url}. Changes:`);
+                console.log(diff);
+                return url;
+              })
+              .catch(function(error) {
+                console.log(error);
+              })
+          };
+        } else {
+          updateDnsRecordDigitalOcean(domain, 'TXT', txtHypercoreWww, "", 300, conf['digitalOceanAccessToken']);
         }
 
         // Pin WWW site to IPFS
-        if (fs.existsSync(dirWww)) {
-          ipfs.add(globSource(dirWww, { recursive: true, followSymlinks: false }), { cidVersion: 1, pin: true, timeout: 60000 })
-            .then(file => file['cid'])
-            .then(cid => {
-              console.log(`WWW site for ${projName} pinned at ipfs/${cid}`);
+        if (protoIpfs) {
+          if (fs.existsSync(dirWww)) {
+            ipfs.add(globSource(dirWww, { recursive: true, followSymlinks: false }), { cidVersion: 1, pin: true, timeout: 60000 })
+              .then(file => file['cid'])
+              .then(cid => {
+                console.log(`WWW site for ${projName} pinned at ipfs/${cid}`);
 
-              // Update DNS record
-              return updateDnsRecordDigitalOcean(domain, 'TXT', txtIpfsWww, `dnslink=/ipfs/${cid}`, 300, conf['digitalOceanAccessToken']);
-            })
-            .catch(function(error) {
-              console.log(error);
-            });
+                // Update DNS record
+                return updateDnsRecordDigitalOcean(domain, 'TXT', txtIpfsWww, `dnslink=/ipfs/${cid}`, 300, conf['digitalOceanAccessToken']);
+              })
+              .catch(function(error) {
+                console.log(error);
+              });
+          }
+        } else {
+          updateDnsRecordDigitalOcean(domain, 'TXT', txtIpfsWww, "", 300, conf['digitalOceanAccessToken']);
         }
 
         // Pin API responses to Hypercore
-        if (fs.existsSync(dirApi)) {
-          getDatSeed('dat-seed-api', projName, domain, txtHypercoreApi)
-            .then(seed => hyperPublisher.sync({
-              seed,
-              fsPath: dirApi,
-              drivePath: '/'
-            }))
-            .then(({diff, url}) => {
-              // Log any changes to hyperdrive and return url
-              console.log(`API responses for ${projName} pinned at ${url}. Changes:`);
-              console.log(diff);
-              return url;
-            })
-            .catch(function(error) {
-              console.log(error);
-            });
+        if (protoHypercore) {
+          if (fs.existsSync(dirApi)) {
+            getDatSeed('dat-seed-api', projName, domain, txtHypercoreApi)
+              .then(seed => hyperPublisher.sync({
+                seed,
+                fsPath: dirApi,
+                drivePath: '/'
+              }))
+              .then(({ diff, url }) => {
+                // Log any changes to hyperdrive and return url
+                console.log(`API responses for ${projName} pinned at ${url}. Changes:`);
+                console.log(diff);
+                return url;
+              })
+              .catch(function(error) {
+                console.log(error);
+              });
+          }
         }
 
         // Pin API responses to IPFS
-        if (fs.existsSync(dirApi)) {
-          ipfs.add(globSource(dirApi, { recursive: true, followSymlinks: false }), { cidVersion: 1, pin: true, timeout: 60000 })
-            .then(file => file['cid'])
-            .then(cid => {
-              console.log(`API responses for ${projName} pinned at ipfs/${cid}`);
+        if (protoIpfs) {
+          if (fs.existsSync(dirApi)) {
+            ipfs.add(globSource(dirApi, { recursive: true, followSymlinks: false }), { cidVersion: 1, pin: true, timeout: 60000 })
+              .then(file => file['cid'])
+              .then(cid => {
+                console.log(`API responses for ${projName} pinned at ipfs/${cid}`);
 
-              // Update DNS record
-              return updateDnsRecordDigitalOcean(domain, 'TXT', txtIpfsApi, `dnslink=/ipfs/${cid}`, 300, conf['digitalOceanAccessToken']);
-            })
-            .catch(function(error) {
-              console.log(error);
-            });
+                // Update DNS record
+                return updateDnsRecordDigitalOcean(domain, 'TXT', txtIpfsApi, `dnslink=/ipfs/${cid}`, 300, conf['digitalOceanAccessToken']);
+              })
+              .catch(function(error) {
+                console.log(error);
+              });
+          }
+        } else {
+          updateDnsRecordDigitalOcean(domain, 'TXT', txtIpfsApi, "", 300, conf['digitalOceanAccessToken']);
+        }
+
+        // Update HTTP A and AAAA record
+        if (protoHttpDnsUpdate) {
+          updateDnsRecordDigitalOcean(domain, 'AAAA', "@", protoHttp ? `${httpPublisherIpv6Address}` : "", 300, conf['digitalOceanAccessToken']);
+          updateDnsRecordDigitalOcean(domain, 'A', "@", protoHttp ? `${httpPublisherIpv4Address}` : "", 300, conf['digitalOceanAccessToken']);
         }
       } catch (error) {
         console.log(error);
@@ -181,16 +231,16 @@ async function getDatSeed(datSeedName, projName, domain, recordName) {
     console.log(`Project ${datSeedName} updated for ${projName}`);
 
     // Publish hyperdrive to dat-store with new seed
-    await hyperPublisher.getURL({seed})
+    await hyperPublisher.getURL({ seed })
       .then(url => datStoreClient.add(url))
-      .then(() => hyperPublisher.create({seed}))
-      .then(({url}) => {
-          console.log(`New hyperdrive published for ${projName}`);
+      .then(() => hyperPublisher.create({ seed }))
+      .then(({ url }) => {
+        console.log(`New hyperdrive published for ${projName}`);
 
-          // Update DNS record
-          const hyperdriveKey = url.replace('hyper://', '');
-          return updateDnsRecordDigitalOcean(domain, 'TXT', recordName, `datkey=${hyperdriveKey}`, 300, conf['digitalOceanAccessToken']);
-        });
+        // Update DNS record
+        const hyperdriveKey = url.replace('hyper://', '');
+        return updateDnsRecordDigitalOcean(domain, 'TXT', recordName, `datkey=${hyperdriveKey}`, 300, conf['digitalOceanAccessToken']);
+      });
   }
   if (seed === undefined) {
     throw new Error(`Failed to get ${datSeedName}`);
@@ -212,24 +262,58 @@ function updateDnsRecordDigitalOcean(domain, recordType, recordName, recordData,
     .then(res => res.json())
     .then(json => json['domain_records'].filter(record => record['type'] === recordType && record['name'] === recordName))
     .then(txt => {
-      // Delete existing records with matching record type and name
-      txt.forEach((item, index) => {
-        const urlDelete = url + item['id'];
-        console.log(`DELETE ${urlDelete}`);
-        fetch(urlDelete, { method: 'DELETE', headers: headers })
-          .catch(function(error) {
-            console.log(error);
-          });
-      });
+      let isDelete = false;
+      let isUpdate = false;
 
-      // Create new DNS record
-      const data = JSON.stringify({
-        type: recordType,
-        name: recordName,
-        data: recordData,
-        ttl:  recordTtl
-      });
-      console.log(`POST ${url} with data ${data}`);
-      return fetch(url, { method: 'POST', headers: headers, body: data });
+      // If there is anything other then one entry found, 
+      // all entries should be removed then one entry added
+      if (txt.length != 1)
+        isUpdate = true
+
+      // If there is a record, and recordData is empty an string, delete only
+      if (txt.length == 1 && recordData == "") {
+        isDelete = true;
+        isUpdate = false
+      }
+
+      // If there is a record, and the data in that record differs from recordData, update
+      if (txt.length == 1 && txt[0]['data'] != recordData)
+        isUpdate = true;
+
+      // If there is no record, and the recordData is an empty string, don't do anything
+      if (txt.length == 0 && recordData == "") {
+        isUpdate = false;
+        isDelete = false;
+      }
+
+      // If there is an update, there needs to be a delete first
+      if (isUpdate == true) isDelete = true;
+
+      // Delete existing records with matching record type and name
+      if (isDelete) {
+        txt.forEach((item, index) => {
+          const urlDelete = url + item['id'];
+          console.log(`DELETE ${urlDelete}`);
+          fetch(urlDelete, { method: 'DELETE', headers: headers })
+            .catch(function(error) {
+              console.log(error);
+            });
+        });
+      }
+
+      // Create new DNS record is we are updating
+      if (isUpdate) {
+
+        const data = JSON.stringify({
+          type: recordType,
+          name: recordName,
+          data: recordData,
+          ttl: recordTtl
+        });
+        console.log(`POST ${url} with data ${data}`);
+        return fetch(url, { method: 'POST', headers: headers, body: data });
+      } else {
+        return;
+      }
     });
 }
